Move SmartStreamInstance and RecordsMemory to smart_stream/mod
