+++ src/spu/src/smart_stream/mod.rs
@@ -1,9 +1,9 @@
 use std::path::Path;
-use std::sync::RwLock;
+use std::sync::{RwLock, Mutex};
 
-use anyhow::Result;
-use tracing::instrument;
-use wasmtime::{Engine, Module, Store};
+use anyhow::{Result, Error};
+use tracing::{debug, instrument};
+use wasmtime::{Engine, Module, Store, Instance, Memory};
 use crate::smart_stream::filter::SmartFilter;
 
 pub mod filter;
@@ -70,5 +70,68 @@ impl SmartStreamModuleInner {
         let store = Store::new(&engine);
         let module = Module::from_binary(store.engine(), binary)?;
         Ok(Self { module, store })
+    }
+}
+
+pub struct SmartStreamInstance(Instance);
+
+impl SmartStreamInstance {
+    fn new(instance: Instance) -> Self {
+        Self(instance)
+    }
+
+    pub fn copy_memory_to(&self, bytes: &[u8]) -> Result<isize, Error> {
+        use self::memory::copy_memory_to_instance;
+        copy_memory_to_instance(bytes, &self.0)
+    }
+}
+
+#[derive(Clone)]
+pub struct RecordsMemory {
+    ptr: i32,
+    len: i32,
+    memory: Memory,
+}
+
+impl RecordsMemory {
+    fn copy_memory_from(&self) -> Vec<u8> {
+        unsafe {
+            if let Some(data) = self
+                .memory
+                .data_unchecked()
+                .get(self.ptr as u32 as usize..)
+                .and_then(|arr| arr.get(..self.len as u32 as usize))
+            {
+                debug!(wasm_mem_len = self.len, "copying from wasm");
+                let mut bytes = vec![0u8; self.len as u32 as usize];
+                bytes.copy_from_slice(data);
+                bytes
+            } else {
+                vec![]
+            }
+        }
+    }
+}
+
+pub struct RecordsCallBack(Mutex<Option<RecordsMemory>>);
+
+impl RecordsCallBack {
+    fn new() -> Self {
+        Self(Mutex::new(None))
+    }
+
+    fn set(&self, records: RecordsMemory) {
+        let mut write_inner = self.0.lock().unwrap();
+        write_inner.replace(records);
+    }
+
+    fn clear(&self) {
+        let mut write_inner = self.0.lock().unwrap();
+        write_inner.take();
+    }
+
+    fn get(&self) -> Option<RecordsMemory> {
+        let reader = self.0.lock().unwrap();
+        reader.clone()
     }
 }
